fix(quizzes): add conditional and auth condi in quiz routes
