+++ apps/web/app/api/room/[roomId]/quizzes/[quizId]/end/route.ts
@@ -1,9 +1,11 @@
 import { prisma } from "@repo/db";
 import { NextRequest, NextResponse } from "next/server";
+import { authMiddleware } from "../../../../../lib/authMiddleware";
 
 export async function PUT(req: NextRequest, { params }: { params: Promise<{roomId: string, quizId: string}>}){
     const { roomId, quizId } = await params;
-
+    const auth = await authMiddleware(req)
+    if(!("authorized" in auth)) return auth;
     try {
         const room = await prisma.room.findUnique({
             where: {
@@ -17,12 +19,18 @@ export async function PUT(req: NextRequest, { params }: { params: Promise<{roomI
             }, {status: 404})
         }
         
-        if (room.endDate < new Date()) {
+        if (room.endDate < new Date() || room.status === "ENDED") {
             return NextResponse.json({
                 message: "Room has already ended. Cannot create quiz."
             }, { status: 400 });
         }
 
+        if(room.creatorId !== auth.userId){
+            return NextResponse.json({
+                message: "Only room owners can end the quiz"
+            }, {status: 403})
+        }
+        
         const quiz = await prisma.quiz.update({
             where: {
                 id: quizId,
