+++ apps/web/app/api/room/[roomId]/quizzes/[quizId]/questions/[questionId]/launch/route.ts
@@ -1,9 +1,11 @@
 import { prisma } from "@repo/db";
 import { NextRequest, NextResponse } from "next/server";
+import { authMiddleware } from "../../../../../../../lib/authMiddleware";
 
 export async function PUT(req: NextRequest, { params }: { params: Promise<{roomId: string, quizId: string, questionId: string}>}){
     const { roomId, quizId, questionId } = await params;
-
+    const auth = await authMiddleware(req)
+    if(!("authorized" in auth)) return auth;
     try {
         const room = await prisma.room.findUnique({
             where: {
@@ -17,10 +19,16 @@ export async function PUT(req: NextRequest, { params }: { params: Promise<{roomI
             }, {status: 404})
         }
         
-        if (room.endDate < new Date()) {
+        if (room.endDate < new Date() || room.status === "ENDED") {
             return NextResponse.json({
                 message: "Room has already ended. Cannot create quiz."
             }, { status: 400 });
+        }
+
+        if(room.creatorId !== auth.userId){
+            return NextResponse.json({
+                message: "Only room owners can launch the question in a quiz"
+            }, {status: 403})
         }
 
         const quiz = await prisma.quiz.findUnique({
@@ -82,74 +90,6 @@ export async function PUT(req: NextRequest, { params }: { params: Promise<{roomI
         }, {status: 200})
     }catch(err) {
         console.log("Error updating the question: ", err);
-        return NextResponse.json({ message: "Internal server error" }, {status: 500})
-    }
-}
-
-export async function GET(req: NextRequest, { params }: { params: Promise<{roomId: string, quizId: string, questionId: string}>}){
-    const { roomId, quizId, questionId } = await params;
-
-    try {
-        const room = await prisma.room.findUnique({
-            where: {
-                id: roomId
-            }
-        })
-
-        if(!room){
-            return NextResponse.json({
-                message: "Room not found"
-            }, {status: 404})
-        }
-        
-        if (room.endDate < new Date()) {
-            return NextResponse.json({
-                message: "Room has already ended. Cannot create quiz."
-            }, { status: 400 });
-        }
-
-        const quiz = await prisma.quiz.findUnique({
-            where: {
-                id: quizId
-            }
-        })
-
-        if(!quiz){
-            return NextResponse.json({
-                message: "Quiz not found"
-            }, {status: 404})
-        }
-
-        const question = await prisma.quizQuestion.findUnique({
-            where: {
-                id: questionId
-            },
-            select: {
-                id: true,
-                quizId: true,
-                question: true,
-                quizOptions: {
-                    select: {
-                        id: true,
-                        text: true,
-                        voteCount: true
-                    }
-                }
-            }
-        })
-
-        if (!question || question.quizId !== quizId) {
-            return NextResponse.json({
-                message: "Question not found in this quiz."
-            }, { status: 404 });
-        }
-        
-        return NextResponse.json({
-            message: "Question fetched successfully",
-            question
-        }, {status: 200})
-    }catch(err) {
-        console.log("Error fetching the question: ", err);
         return NextResponse.json({ message: "Internal server error" }, {status: 500})
     }
 }
